requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
django-cors-headers>=4.3
django-simple-history>=3.4
pillow>=10.2
pillow-heif>=0.12
requests>=2.31
kavenegar>=1.1
qrcode>=7.4
reportlab>=4.1
arabic-reshaper>=3.0
python-bidi>=0.4
jdatetime>=4.1
langchain>=0.1
langchain-core>=0.1
langchain-community>=0.0.21
langchain-text-splitters>=0.0.1
freezegun>=1.2
pytest>=8.1
pytest-cov>=4.1
pytest-django>=4.8
gunicorn>=21.2
mysqlclient>=2.2.7,<3